Update Feature_Test.java
Alright so I have done a few changes. I have done some changes to the menu so that it looks neat. Also for the menu I added a Quit Button and a Default case. I also put my Randomized board into its own method and made that method void so that it outputs the board. We can create another method to do the actual moves.